make sure recovery partition flag is checked
will check for: TW_HAS_RECOVERY_PARTITION which is set (or not) in BoardConfig.mk
when unset there or set to "false" it will result in a ramdisk cpio.
when set to "true" it will create a full recovery.img.

just for info:
TW_HAS_RECOVERY_PARTITION will become SHRP_HAS_RECOVERY_PARTITION during the
check process

Signed-off-by: steadfasterX <steadfasterX@gmail.com>